+++ handlers/react_add.go
@@ -1,0 +1,11 @@
+package handlers
+
+import (
+	"github.com/bwmarrin/discordgo"
+)
+
+func ReactAdd(s *discordgo.Session, r *discordgo.MessageReactionAdd) {
+	if r.UserID == s.State.User.ID {
+		return
+	}
+}